Moved custom/utils/format to cow-react/utils/format
