+++ src/cow-react/modules/limitOrders/containers/LimitOrdersConfirmModal/index.tsx
@@ -11,7 +11,7 @@ import { limitOrdersConfirmState } from '../LimitOrdersConfirmModal/state'
 import { useWalletInfo } from 'hooks/useWalletInfo'
 import { GpModal } from 'components/Modal'
 import * as styledEl from './styled'
-import { formatSmartAmount } from 'utils/format'
+import { formatSmartAmount } from '@cow/utils/format'
 import { useRateImpact } from '@cow/modules/limitOrders/hooks/useRateImpact'
 import { useRateInfoParams } from '@cow/common/hooks/useRateInfoParams'
 import { LimitOrdersWarnings } from '@cow/modules/limitOrders/containers/LimitOrdersWarnings'
