+++ src/cow-react/modules/swap/containers/Row/RowSlippage/index.tsx
@@ -3,7 +3,7 @@ import { Percent } from '@uniswap/sdk-core'
 
 import { PERCENTAGE_PRECISION } from 'constants/index'
 import { useToggleSettingsMenu } from 'state/application/hooks'
-import { formatSmart } from 'utils/format'
+import { formatSmart } from '@cow/utils/format'
 import { RowSlippageContent } from '@cow/modules/swap/pure/Row/RowSlippageContent'
 import { useIsEthFlow } from '@cow/modules/swap/hooks/useIsEthFlow'
 import { useDetectNativeToken } from '@cow/modules/swap/hooks/useDetectNativeToken'
