+++ src/cow-react/modules/swap/helpers/tradeReceiveAmount.ts
@@ -1,5 +1,5 @@
 import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
-import { formatSmartAmount } from 'utils/format'
+import { formatSmartAmount } from '@cow/utils/format'
 import TradeGp from 'state/swap/TradeGp'
 
 export interface ReceiveAmountInfo {
