+++ src/cow-react/modules/swap/pure/EthFlow/WrappingPreview/WrapCard.tsx
@@ -1,7 +1,7 @@
 import { CurrencyAmount, Currency } from '@uniswap/sdk-core'
 import styled from 'styled-components/macro'
 import CurrencyLogo from 'components/CurrencyLogo'
-import { formatSmartAmount } from 'utils/format'
+import { formatSmartAmount } from '@cow/utils/format'
 import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
 import { CHAIN_INFO } from 'constants/chainInfo'
 import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
