+++ src/cow-react/pages/Account/LockedGnoVesting/index.tsx
@@ -7,7 +7,7 @@ import { MouseoverTooltipContent } from 'components/Tooltip'
 import cowImage from 'assets/cow-swap/cow_v2.svg'
 import ArrowIcon from 'assets/cow-swap/arrow.svg'
 import { AMOUNT_PRECISION } from 'constants/index'
-import { formatSmartLocaleAware } from 'utils/format'
+import { formatSmartLocaleAware } from '@cow/utils/format'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { useErrorModal } from 'hooks/useErrorMessageAndModal'
 import CopyHelper from 'components/Copy'
