+++ src/cow-react/pages/Account/utils.ts
@@ -1,4 +1,4 @@
-import { numberFormatter } from 'utils/format'
+import { numberFormatter } from '@cow/utils/format'
 
 export const formatDecimal = (number?: number): string => {
   return number ? numberFormatter.format(number) : '-'
