+++ src/cow-react/pages/Claim/InvestmentFlow/InvestSummaryRow.tsx
@@ -3,7 +3,7 @@ import { calculatePercentage } from 'state/claim/hooks/utils'
 import { TokenLogo, InvestAvailableBar, UserMessage } from '@cow/pages/Claim/styled'
 import { ClaimWithInvestmentData } from '@cow/pages/Claim/types'
 import CowProtocolLogo from 'components/CowProtocolLogo'
-import { formatMax, formatSmartLocaleAware } from 'utils/format'
+import { formatMax, formatSmartLocaleAware } from '@cow/utils/format'
 import { ONE_HUNDRED_PERCENT } from 'constants/misc'
 import { AMOUNT_PRECISION } from 'constants/index'
 import ImportantIcon from 'assets/cow-swap/important.svg'
