+++ src/custom/components/CowBalanceButton/index.tsx
@@ -3,7 +3,7 @@ import styled, { css } from 'styled-components/macro'
 import CowProtocolLogo from 'components/CowProtocolLogo'
 import { useCombinedBalance } from 'state/cowToken/hooks'
 import { ChainId } from 'state/lists/actions/actionsMod'
-import { formatMax, formatSmartLocaleAware } from 'utils/format'
+import { formatMax, formatSmartLocaleAware } from '@cow/utils/format'
 import { COW } from 'constants/tokens'
 import { transparentize } from 'polished'
 import { useWeb3React } from '@web3-react/core'
