+++ src/custom/components/Tokens/FiatBalanceCell.tsx
@@ -3,7 +3,7 @@ import { Token, CurrencyAmount } from '@uniswap/sdk-core'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
 import { FIAT_PRECISION } from 'constants/index'
 import { BalanceValue, InfoCircle, FiatValue } from './styled'
-import { formatMax, formatSmart } from 'utils/format'
+import { formatMax, formatSmart } from '@cow/utils/format'
 import { MouseoverTooltip } from 'components/Tooltip'
 
 type FiatBalanceCellProps = {
