+++ src/custom/utils/trade.ts
@@ -8,7 +8,7 @@ import { OrderID, sendOrder as sendOrderApi, sendSignedOrderCancellation } from 
 import { Signer } from '@ethersproject/abstract-signer'
 import { AMOUNT_PRECISION, RADIX_DECIMAL, NATIVE_CURRENCY_BUY_ADDRESS } from 'constants/index'
 import { SupportedChainId as ChainId } from 'constants/chains'
-import { formatSmart } from 'utils/format'
+import { formatSmart } from '@cow/utils/format'
 import { SigningScheme } from '@cowprotocol/contracts'
 import { getProfileData, getTrades } from '@cow/api/gnosisProtocol/api'
